+++ src/checkovRunner.ts
@@ -1,4 +1,6 @@
-import { spawn } from "child_process";
+import * as vscode from 'vscode';
+import { spawn } from 'child_process';
+import { Logger } from 'winston';
 
 export interface FailedCheckovCheck {
     checkId: string;
@@ -30,32 +32,40 @@ interface CheckovResponseRaw {
     };
 }
 
-export const runCheckovScan = (fileName: string): Promise<CheckovResponse> => {
+const skipChecks = ['CKV_AWS_52'];
+
+export const runCheckovScan = (logger: Logger, fileName: string, token: string, cancelToken: vscode.CancellationToken): Promise<CheckovResponse> => {
     return new Promise((resolve, reject) => {
-        console.log('Running checkov on', fileName);
-        const ckv = spawn('checkov', ['-f', fileName, '-o', 'json']);
+        logger.info('Running checkov on', { fileName });
+        const ckv = spawn('checkov', ['-s', '--skip-check', skipChecks.join(','), '--bc-api-key', token, '--repo-id', 'vscode/extension', '-f', fileName, '-o', 'json']);
         let stdout = '';
 	
-        ckv.stdout.on("data", data => {
+        ckv.stdout.on('data', data => {
             stdout += data;
         });
 			
-        ckv.stderr.on("data", data => {
-            console.warn(`Checkov stderr: ${data}`);
+        ckv.stderr.on('data', data => {
+            logger.warn(`Checkov stderr: ${data}`);
         });
 			
         ckv.on('error', (error) => {
-            console.error('Error while running Checkov', error);
+            logger.error('Error while running Checkov', { error });
         });
 			
-        ckv.on("close", code => {
-            console.debug(`Checkov scan process exited with code ${code}`);
-            if (code !== 1) return reject(`Checkov exited with code ${code}`); // Check about checkov
-	
-            console.debug(`Checkov task output: ${stdout}`);
+        ckv.on('close', code => {
+            if (cancelToken.isCancellationRequested) return reject('Cancel invoked');
+            logger.debug(`Checkov scan process exited with code ${code}`);
+            if (code !== 0) return reject(`Checkov exited with code ${code}`);
+            
             const output: CheckovResponseRaw = JSON.parse(stdout);
+            logger.debug('Checkov task output:', output);
 	
             resolve(parseCheckovResponse(output));
+        });
+
+        cancelToken.onCancellationRequested((cancelEvent) => {
+            ckv.kill('SIGABRT');
+            logger.info('Cancellation token invoked, aborting checkov run', { cancelEvent });
         });
     });
 };