+++ src/utils.ts
@@ -1,7 +1,9 @@
 import * as vscode from 'vscode';
 import { exec } from 'child_process';
+import winston = require('winston');
 import { FailedCheckovCheck } from './checkovRunner';
 import { DiagnosticReferenceCode } from './diagnostics';
+import { CHECKOV_MAP } from './extension';
 
 type ExecOutput = [stdout: string, stderr: string];
 export const asyncExec = async (commandToExecute: string) : Promise<ExecOutput> => {
@@ -13,6 +15,32 @@ export const asyncExec = async (commandToExecute: string) : Promise<ExecOutput> 
     });
 }; 
 
+export const saveCheckovResult = (state: vscode.Memento, checkovFails: FailedCheckovCheck[]): void => {
+    const checkovMap = checkovFails.reduce((prev, current) => ({
+        ...prev,
+        [createCheckovKey(current)]: current
+    }), []);
+    state.update(CHECKOV_MAP, checkovMap);
+};
+
 export const createDiagnosticKey = (diagnostic: vscode.Diagnostic): string => 
     `${(diagnostic.code as DiagnosticReferenceCode).value}-${diagnostic.range.start.line + 1}`;
 export const createCheckovKey = (checkovFail: FailedCheckovCheck): string => `${checkovFail.checkId}-${checkovFail.fileLineRange[0]}`;
+
+export const getLogger = (logFileDir: string, logFileName: string): winston.Logger => winston.createLogger({
+    level: 'debug',
+    format: winston.format.combine(
+        winston.format.splat(),
+        winston.format.printf(({ level, message, ...rest }) => {
+            const logError = rest.error ? { error: { ...rest.error, message: rest.error.message, stack: rest.error.stack } } : {};
+            return `[${level}]: ${message} ${JSON.stringify({ ...rest, ...logError })}`; 
+        })
+    ),
+    transports: [
+        new winston.transports.File({
+            level: 'debug',
+            dirname: logFileDir,
+            filename: logFileName
+        })
+    ]
+});